% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seq_id_extras.R
\name{add_seq_id_extras}
\alias{add_seq_id_extras}
\title{add_seq_id_extras function}
\usage{
add_seq_id_extras(df, species)
}
\arguments{
\item{df}{A data frame of one species}

\item{species}{A string that represent a species name}
}
\value{
A data frame with seq_id column added
}
\description{
Add seq_id column to extras fragments tables
}
\details{
Belong to the pipeline function in step 11 : create aligned sequences, aligned genes and links tables for gggenomes
}
\examples{
#...
}
