% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligned_sequences_extension.R
\name{aligned_sequences_extension}
\alias{aligned_sequences_extension}
\title{aligned_sequences_extension function}
\usage{
aligned_sequences_extension(aligned_sequences, seqs)
}
\arguments{
\item{aligned_sequences}{An aligned sequences data frame to be display}

\item{seqs}{An aligned sequences used in gggenomes}
}
\value{
An updated aligned sequences data frame to be display
}
\description{
Update aligned sequences table to be display
}
\details{
Belong to the server function in the full genes extension button section
}
\examples{
#...
}
