% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_start.R
\name{calculate_start}
\alias{calculate_start}
\title{calculate_start function}
\usage{
calculate_start(df)
}
\arguments{
\item{df}{A data frame issued from halLiftover linux command}
}
\value{
A data frame with start column calculated
}
\description{
Calculate the start column of a PSL file
}
\details{
Belong to the sort_halLiftover_data function in the pipeline function of the step 1 : retrieve the homologous fragments of the C. Elegans sequence of inquiry
}
\examples{
#...
}
