% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overlapping.R
\name{check_overlapping}
\alias{check_overlapping}
\title{check_overlapping function}
\usage{
check_overlapping(species, df)
}
\arguments{
\item{species}{A species specific data frame}

\item{df}{An aligned genes data frame}
}
\value{
An updated aligned genes data frame
}
\description{
Overlap genes
}
\details{
Belong to the pipeline function in step 8 : check if there is some hidden genes inside the intersected genes found
}
\examples{
#...
}
