% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_strand.R
\name{convert_strand}
\alias{convert_strand}
\title{convert_strand function}
\usage{
convert_strand(df)
}
\arguments{
\item{df}{A data frame from HalLiftover with a strand column}
}
\value{
A data frame with two strands columns (query and target)
}
\description{
Convert tEnds and tStarts columns if the strand negative
}
\details{
Belong to the sort_halLiftover_data function of the pipeline function in step 1 : retrieve the homologous fragments of the C. Elegans sequence of inquiry
}
\examples{
#...
}
