% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_infos.R
\name{display_infos}
\alias{display_infos}
\title{display_infos function}
\usage{
display_infos(genes, aligned_sequences, links, region)
}
\arguments{
\item{genes}{The aligned genes data frame}

\item{aligned_sequences}{The aligned sequences data frame}

\item{links}{The links data frame}

\item{region}{The region data frame}
}
\value{
A list containing the genes, the aligned sequences, the links and the region data frames
}
\description{
Display information nicely in the UI table panels
}
\details{
Belong to the server function in every sections
}
\examples{
#...
}
