% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_fragments.R
\name{extras_fragments}
\alias{extras_fragments}
\title{extras_fragments function}
\usage{
extras_fragments(extras_fragments, top_fragments)
}
\arguments{
\item{extras_fragments}{A data frame containing the extras fragments}

\item{top_fragments}{A data frame containing all fragments}
}
\value{
A data frame of extras fragments
}
\description{
Take extras_fragments from top fragments tables
}
\details{
Belong to the pipeline function in step 5 : separate the longest homologous fragments from the others
}
\examples{
#...
}
