% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_seqs_fct.R
\name{extras_seqs_fct}
\alias{extras_seqs_fct}
\title{extras_seqs_fct function}
\usage{
extras_seqs_fct(extras_genes, extras_fragments, species)
}
\arguments{
\item{extras_genes}{A data frame of extras genes}

\item{extras_fragments}{A data frame of extras fragments}

\item{species}{A species specific data frame}
}
\value{
A data frame of extras sequences
}
\description{
Create a data frame of extras sequences
}
\details{
Belong to the pipeline function in step 11 : create aligned sequences, aligned genes and links tables for gggenomes
}
\examples{
#...
}
