% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{filtering function}
\usage{
filtering(stitch_data, region)
}
\arguments{
\item{stitch_data}{A data frame of assembled fragments}

\item{region}{The region inquiry data frame}
}
\value{
A data frame of filtered assembled fragments
}
\description{
Filter assembled fragments according to a filtering value
}
\details{
Belong to the pipeline function in step 4 : filter assembled homologous fragments through a filtering percentage of the sequence of inquiry length
}
\examples{
#...
}
