% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{final}
\alias{final}
\title{final function}
\usage{
final(filtered_data)
}
\arguments{
\item{filtered_data}{A data frame of the assembled fragments}
}
\value{
A data frame of assembled fragments
}
\description{
Keep only the longest assembled fragment
}
\details{
Belong to the pipeline function in step 5 : separate the longest homologous fragments from the others
}
\examples{
#...
}
