% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_genes_fct.R
\name{full_genes_fct}
\alias{full_genes_fct}
\title{full_genes_fct function}
\usage{
full_genes_fct(genes, species)
}
\arguments{
\item{genes}{A data frame of aligned genes}

\item{species}{A data frame of genes of a certain species}
}
\value{
A data frame of extended aligned genes
}
\description{
Extend genes to their full length
}
\details{
Belong to the server function in the full genes extension section
}
\examples{
#...
}
