% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_name_fct.R
\name{gene_name_fct}
\alias{gene_name_fct}
\title{gene_name_fct function}
\usage{
gene_name_fct(genes, region, elegans)
}
\arguments{
\item{genes}{A data frame of aligned genes}

\item{region}{The region inquiry data frame}

\item{elegans}{The C. Elegans genes data frame}
}
\value{
A data frame of aligned genes with C. Elegans genes names
}
\description{
Create a table with C. Elegans genes names, orthogroup and color information
}
\details{
Belong to the plot_2 and plot_3 functions in the ggg_plot function from the server function in every section
}
\examples{
#...
}
