% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_species.R
\name{grab_species}
\alias{grab_species}
\title{grab_species function}
\usage{
grab_species(genes, seqs)
}
\arguments{
\item{genes}{An aligned genes data frame}

\item{seqs}{An aligned sequences data frame used in gggenomes}
}
\value{
A list of two vectors of missing genes and missing fragments
}
\description{
Make a list of species that do not have aligned regions and genes
}
\details{
Belong to the render_explanations function in the server function in every section
}
\examples{
#...
}
