% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_fct.R
\name{hide_fct}
\alias{hide_fct}
\title{hide_fct function}
\usage{
hide_fct(seqs, aligned_sequences, genes)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{aligned_sequences}{An aligned sequences data frame to be diplay}

\item{genes}{An aligned genes data frame}
}
\value{
A list of data frames : aligned sequences, seqs, and genes
}
\description{
Hide the extras fragments in the aligned genes, aligned sequences and links tables
}
\details{
Belong to the server function in the hide button section
}
\examples{
#...
}
