% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{history}
\alias{history}
\title{history function}
\usage{
history(
  filesHistory,
  plot,
  genes,
  seqs,
  aligned_sequences,
  links,
  region,
  prev_colors,
  modifications
)
}
\arguments{
\item{filesHistory}{A list of lists}

\item{plot}{A gggenomes generated plot}

\item{genes}{An aligned genes data frame}

\item{seqs}{An aligned sequences data frame to be display}

\item{aligned_sequences}{An aligned sequences data frame used in gggenomes}

\item{links}{A links data frame}

\item{region}{A region inquiry data frame}

\item{prev_colors}{A vector of orthogroups and colors}

\item{modifications}{A string characterizing an app action}
}
\value{
An updated History list
}
\description{
Update the History list with the used tables, plot, string and vector
}
\details{
Belong to the server function in every section
}
\examples{
#...
}
