% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links_fct.R
\name{links_fct}
\alias{links_fct}
\title{links_fct function}
\usage{
links_fct(species1_species2)
}
\arguments{
\item{species1_species2}{A data frame of links}
}
\value{
A data frame of links
}
\description{
Select the columns that match the requirements of links table in gggenomes
}
\details{
Belong to the pipeline function in step 2 : find pairwise alignment between phylogenetically adjacent species using the homologous fragments
}
\examples{
#...
}
