% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesSelectize.R
\name{load_genesSelectize}
\alias{load_genesSelectize}
\title{load_genesSelectize function}
\usage{
load_genesSelectize(session)
}
\arguments{
\item{session}{A shiny app session}
}
\value{
An updated UI with the list of genes in the Coordinates panel
}
\description{
Load the genes list in the Coordinates panel
}
\details{
Belong to the server function independent of a section
}
\examples{
#...
}
