% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{macro}
\alias{macro}
\title{macro function}
\usage{
macro(genes, seqs, species)
}
\arguments{
\item{genes}{An aligned genes data frame}

\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{species}{A species specific data frame}
}
\value{
An aligned genes data frame with gene gene type
}
\description{
Transform aligned genes data to macrosynteny data
}
\details{
Belong to the server function in the macrosynteny button section
}
\examples{
#...
}
