% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messages}
\alias{messages}
\title{messages function}
\usage{
messages(regions, genes)
}
\arguments{
\item{regions}{A vector of species that do not have fragments}

\item{genes}{A vector of species that do not have genes}
}
\value{
A string : the messages shown in explanation panel
}
\description{
Create messages in functions of missing fragments and genes
}
\details{
Belong to the render_explanations function in the server function in every section
}
\examples{
#...
}
