% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_fct.R
\name{metadata_fct}
\alias{metadata_fct}
\title{metadata_fct function}
\usage{
metadata_fct(species, genes_species)
}
\arguments{
\item{species}{A specific species data frame}

\item{genes_species}{An aligned genes data frame}
}
\value{
An aligned genes data frame that contains metadata
}
\description{
Add metadata to the aligned genes data frames
}
\details{
Belong to the pipeline function in step 7 : add metadata to genes/piece of genes in the aligned regions and step 9 : find genes/piece of genes in the removed regions too
}
\examples{
#...
}
