% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1.R
\name{plot_1}
\alias{plot_1}
\title{plot_1 function}
\usage{
plot_1(seqs, region)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{region}{A region inquiry data frame}
}
\value{
A gggenomes plot
}
\description{
Generate plot case 1 : no genes and links
}
\details{
Belong to ggg_plot function in the server function in every section
}
\examples{
#...
}
