% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2.R
\name{plot_2}
\alias{plot_2}
\title{plot_2 function}
\usage{
plot_2(seqs, genes, region, elegans_file)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{genes}{An aligned genes data frame}

\item{region}{A region inquiry data frame}

\item{elegans_file}{A data frame containg C. Elegans genes names of genes table}
}
\value{
A gggenomes plot
}
\description{
Generate plot case 2 : no links
}
\details{
Belong to the ggg_plot function in the server function in every section
}
\examples{
#...
}
