% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3.R
\name{plot_3}
\alias{plot_3}
\title{plot_3 function}
\usage{
plot_3(seqs, genes, links, region, elegans_file)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{genes}{An aligned genes data frame}

\item{links}{A links data frame}

\item{region}{A region inquiry data frame}

\item{elegans_file}{A data frame containg C. Elegans genes names of genes table}
}
\value{
A gggenomes plot
}
\description{
Generate plot case 3 : everything is present
}
\details{
Belong to the ggg_plot function in the server function in every section
}
\examples{
#...
}
