% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_4.R
\name{plot_4}
\alias{plot_4}
\title{plot_4 function}
\usage{
plot_4(seqs, genes, links, region)
}
\arguments{
\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{genes}{An aligned genes data frame}

\item{links}{A links data frame}

\item{region}{A region inquiry data frame}
}
\value{
A gggenomes plot
}
\description{
Generate plot case 4 : no C. Elegans genes
}
\details{
Belong to the ggg_plot function in the server function in every section
}
\examples{
#...
}
