% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pslToDataFrame.R
\name{pslToDataFrame}
\alias{pslToDataFrame}
\title{pslToDataFrame function}
\usage{
pslToDataFrame(elegans_species, species_name)
}
\arguments{
\item{elegans_species}{A PSL format file}

\item{species_name}{A string that represents a species name}
}
\value{
A data frame of aligned fragments
}
\description{
From PSL file format to data frame format case 1 : C. Elegans as query
}
\details{
Belong to the pipeline function in step 1 : retrieve the homologous fragments of the C. elegans sequence of inquiry
}
\examples{
#...
}
