% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_others.R
\name{render_others}
\alias{render_others}
\title{render_others function}
\usage{
render_others(output, region, input)
}
\arguments{
\item{output}{An ouput from a shiny app session}

\item{region}{A region inquiry data frame}

\item{input}{An input from a shiny app session}
}
\value{
Display the region inquiry information on the explanation panel
}
\description{
Render the region inquiry information on the explanation panel
}
\details{
Belong to the server function in every section
}
\examples{
#...
}
