% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetCoordinates.R
\name{resetCoordinates}
\alias{resetCoordinates}
\title{resetCoordinates function}
\usage{
resetCoordinates(session)
}
\arguments{
\item{session}{A shiny app session}
}
\value{
Coordinates updated on the Coordinates panel
}
\description{
Reset coordinates with default coordinates
}
\details{
Belong to the server function in the reset button section
}
\examples{
#...
}
