% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_same_genes.R
\name{retrieve_same_genes}
\alias{retrieve_same_genes}
\title{retrieve_same_genes function}
\usage{
retrieve_same_genes(genes_file, extras_genes)
}
\arguments{
\item{genes_file}{A top genes data frame}

\item{extras_genes}{An extras genes data frame}
}
\value{
An extras genes data frame
}
\description{
Retrieve the same genes in the extras fragments and top fragments
}
\details{
Belong to the pipeline function in step 10 : remove same genes from extras fragments and place them into top fragments
}
\examples{
#...
}
