% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs_fct.R
\name{seqs_fct}
\alias{seqs_fct}
\title{seqs_fct function}
\usage{
seqs_fct(genes, final_data, species)
}
\arguments{
\item{genes}{An aligned genes data frame}

\item{final_data}{A data frame of aligned fragments}

\item{species}{A specific species data frame}
}
\value{
A formatting data frame of aligned sequences
}
\description{
Format the aligned sequences data frame :  add bin_id
}
\details{
Belong to the pipeline function in step 11 : create aligned sequences, aligned genes and links tables for gggenomes
}
\examples{
#
}
