% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{A Shiny app input session}

\item{output}{A Shiny app output session}

\item{session}{A Shiny app session}
}
\value{
The server part of a Shiny R session
}
\description{
Server function
}
\examples{
#
}
