% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fct.R
\name{show_fct}
\alias{show_fct}
\title{show_fct function}
\usage{
show_fct(genes, seqs, aligned_sequences, selected_species)
}
\arguments{
\item{genes}{An aligned genes data frame}

\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{aligned_sequences}{An aligned sequences data frame to display}

\item{selected_species}{A list of species names}
}
\value{
A list of data frames
}
\description{
Update aligned genes, aligned_sequences (displayed), sequences (gggenomes) and links tables accroding to the user species selection
}
\details{
Belong to the server function in the show button section
}
\examples{
#...
}
