% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_fct.R
\name{switch_fct}
\alias{switch_fct}
\title{switch_fct function}
\usage{
switch_fct(list_species, seqs, genes, aligned_sequences)
}
\arguments{
\item{list_species}{A vector of string that represent species name selected by the user}

\item{seqs}{An aligned sequences data frame used in gggenomes}

\item{genes}{An aligned genes data frame}

\item{aligned_sequences}{An aligned sequences data frame to display}
}
\value{
A list of two data frames : an aligned gene
s data frame and an aligned sequences data frame used in gggenomes
}
\description{
Switch button : switch extras fragments from bottom to top and vice versa
}
\details{
Belong to the server function in the switch button section
}
\examples{
#...
}
