% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_linux.R
\name{system_linux}
\alias{system_linux}
\title{system_linux function}
\usage{
system_linux(species)
}
\arguments{
\item{species}{A vector of string that represent species names}
}
\value{
A list of data frames under the PSL format
}
\description{
Retrieve the homologous fragments of the C. Elegans sequence of inquiry with the halLiftover linux command
}
\details{
Belong to the pipeline function in step 1 : retrieve the homologous fragments of the C. Elegans sequence of inquiry
}
\examples{
#...

}
