% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_linux_2.R
\name{system_linux_2}
\alias{system_linux_2}
\title{system_linux_2 function}
\usage{
system_linux_2(species)
}
\arguments{
\item{species}{A vector of string that represent species names}
}
\value{
A list of data frames under the PSL format
}
\description{
Find pairwise alignment between phylogenetically adjacent species using the halLiftover linux command
}
\details{
Belong to the pipeline function in step 2 : find pairwise alignment between phylogenetically adjacent species using the homologous fragments
}
\examples{
#...
}
