% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_retrieved.R
\name{transform_retrieved}
\alias{transform_retrieved}
\title{transform_retrieved function}
\usage{
transform_retrieved(df, species)
}
\arguments{
\item{df}{A data frame of same genes between extras and top fragments}

\item{species}{A specific species data frame}
}
\value{
A data frame of retrieved genes with seq_id column added
}
\description{
Add seq_id in retrieved genes tables
}
\details{
Belong to the pipeline function in step 10 : remove same genes from extras fragments and place them into top fragments
}
\examples{
#...
}
