% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sequences_table.R
\name{update_sequences_table}
\alias{update_sequences_table}
\title{update_sequences_table function}
\usage{
update_sequences_table(list_species, seqs)
}
\arguments{
\item{list_species}{A list of string that represents species names}

\item{seqs}{An aligned sequences data frame to be display}
}
\value{
An updated aligned sequences data frame to be display
}
\description{
Update the aligned sequences table to be display when switching genes
}
\details{
Belong to the server function in the switch button section
}
\examples{
#...
}
