% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCoordinates.R
\name{updateCoordinates}
\alias{updateCoordinates}
\title{updateCoordinates function}
\usage{
updateCoordinates(session, region)
}
\arguments{
\item{session}{A shiny app session}

\item{region}{A region inquiry data frame}
}
\value{
Updated Coordinates panel
}
\description{
Update coordinates in the Coordinates panel when the user is selecting a gene
}
\details{
Belong to the server function out of a section
}
\examples{
#...
}
