% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_in.R
\name{zoom_in}
\alias{zoom_in}
\title{zoom_in function}
\usage{
zoom_in(region)
}
\arguments{
\item{region}{A region inquiry data frame}
}
\value{
A region inquiry data frame with modified coordinates
}
\description{
Change the region inquiry coordinates
}
\details{
Belong to the server function in the zoom in button section
}
\examples{
#...
}
