% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_out.R
\name{zoom_out}
\alias{zoom_out}
\title{zoom_out function}
\usage{
zoom_out(region)
}
\arguments{
\item{region}{A region inquiry data frame}
}
\value{
A region inquiry data frame with modified coordinates
}
\description{
Change the region inquiry coordinates
}
\details{
Belong to the server function in the zoom out button section
}
\examples{
#...
}
